+++ backend/src/test/java/com/example/catalog/product/service/ProductServiceTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
+import org.mockito.junit.jupiter.MockitoExtension;\nimport java.math.BigDecimal;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -26,7 +26,7 @@ class ProductServiceTest {
     @InjectMocks
     private ProductService productService;
 
-    private Product createProduct(Long id, String name, String category, String description, double price) {
+    private Product createProduct(Long id, String name, String category, String description, BigDecimal price) {
         Product p = new Product(name, description, price, category);
         p.setId(id);
         return p;
@@ -34,9 +34,9 @@ class ProductServiceTest {
 
     @Test
     void getSimilarProducts_shouldReturnSimilarProducts() {
-        Product product1 = createProduct(1L, "Apple", "Fruit", "A red fruit", 1.0);
-        Product product2 = createProduct(2L, "Banana", "Fruit", "A yellow fruit", 0.5);
-        Product product3 = createProduct(3L, "Carrot", "Vegetable", "An orange vegetable", 0.3);
+        Product product1 = createProduct(1L, "Apple", "Fruit", "A red fruit", BigDecimal.valueOf(1.0));
+        Product product2 = createProduct(2L, "Banana", "Fruit", "A yellow fruit", BigDecimal.valueOf(0.5));
+        Product product3 = createProduct(3L, "Carrot", "Vegetable", "An orange vegetable", BigDecimal.valueOf(0.3));
 
         when(productRepository.findById(1L)).thenReturn(Optional.of(product1));
         when(productRepository.findByCategoryAndIdNot("Fruit", 1L)).thenReturn(Arrays.asList(product2));
@@ -59,7 +59,7 @@ class ProductServiceTest {
 
     @Test
     void getSimilarProducts_shouldReturnEmptyList_whenNoSimilarProducts() {
-        Product product1 = createProduct(1L, "UniqueProduct", "UniqueType", "A very unique description", 1.0);
+        Product product1 = createProduct(1L, "UniqueProduct", "UniqueType", "A very unique description", BigDecimal.valueOf(1.0));
 
         when(productRepository.findById(1L)).thenReturn(Optional.of(product1));
         when(productRepository.findByCategoryAndIdNot("UniqueType", 1L)).thenReturn(Collections.emptyList());
